added leftflipper drawing and cell occupation checking (COLLISIONS STILL NEED DONE)

+++ Gizmoball Code/src/Model/gizmos/SquareBumper.java
@@ -37,12 +37,12 @@ public class SquareBumper implements Gizmo{
     }
 
     @Override
-    public int getWidth() {
+    public float getWidth() {
         return 1;
     }
 
     @Override
-    public int getHeight() {
+    public float getHeight() {
         return 1;
     }
 
@@ -134,7 +134,7 @@ public class SquareBumper implements Gizmo{
     public int[][] getBounds() {
         int[][] bounds = new int[][]{
                 {xPos, yPos},
-                {xPos+getWidth(), yPos+getHeight()}
+                {xPos+(int)getWidth(), yPos+(int)getHeight()}
         };
 
         return bounds;
